+++ src/App.vue
@@ -1,22 +1,45 @@
 <template>
-  <div id="app">
-    <div class="app__container">
-      <div class="app__main-content">
-        <div class="app__navbar"/>
-        <div class="app__main">
-          <div class="app__main-sidebar"/>
-          <div class="app__content-wrapper">
-            <router-view />
-          </div>
-        </div>
+  <div class="app__container">
+    <div class="app__main-content">
+      <div class="app__navbar"/>
+      <div class="app__main">
+        <div class="app__main-sidebar"/>
+        <router-view />
       </div>
     </div>
   </div>
-  <router-view/>
+
+  <status-message />
 </template>
 
 <script>
+import StatusMessage from '@/vue/common/StatusMessage'
+
 export default {
   name: 'app',
+
+  components: { StatusMessage },
 }
 </script>
+
+<style lang="scss" scoped>
+.app__container {
+  display: flex;
+  align-items: stretch;
+  overflow-x: hidden;
+  flex: 1;
+}
+
+.app__main-content {
+  flex: 1;
+  overflow: hidden;
+}
+
+.app__main {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+}
+</style>