+++ src/js/const/user-roles.const.js
@@ -1,0 +1,5 @@
+export const USER_ROLES = Object.freeze({
+  general: 0,
+  notary: 1,
+  registry: 2,
+})